Removed the maps folder and move the map related components into components folder
